% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_functions.R
\name{aggregate_annot}
\alias{aggregate_annot}
\title{Aggregate annotation datasets}
\usage{
aggregate_annot(data_annot, type_annot)
}
\arguments{
\item{data_annot}{???}

\item{type_annot}{???}
}
\description{
Aggregate annotation datasets
}
