% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{common_neighbors}
\alias{common_neighbors}
\title{Shared neighbors between proteins}
\usage{
common_neighbors(graph_op)
}
\arguments{
\item{graph_op}{???}
}
\value{
???
}
\description{
Shared neighbors between proteins
}
