% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_functions.R
\name{count_genes}
\alias{count_genes}
\title{Number of total genes in an ontology database}
\usage{
count_genes(data_annot)
}
\arguments{
\item{data_annot}{Data frame (tibble) of GO or HPO datasets from
\code{\link{wppi_data}}.}
}
\value{
Number of total unique genes in each ontology database.
}
\description{
Number of total genes in an ontology database
}
\examples{
go <- wppi_go_data()
count_genes(go)
# [1] 19712

}
\seealso{
\itemize{
    \item{\code{\link{wppi_data}}}
    \item{\code{\link{wppi_go_data}}}
    \item{\code{\link{wppi_hpo_data}}}
}
}
