% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_functions.R
\name{filter_annot_with_network}
\alias{filter_annot_with_network}
\title{Filter annotation datasets using network object}
\usage{
filter_annot_with_network(data_annot, graph_op)
}
\arguments{
\item{data_annot}{???}

\item{graph_op}{???}
}
\description{
Filter annotation datasets using network object
}
