% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_functions.R
\name{functional_annot}
\alias{functional_annot}
\title{Functional similarity between two genes in annotation database}
\usage{
functional_annot(data_aggregated, nr_genes, gene_i, gene_j)
}
\arguments{
\item{data_aggregated}{???}

\item{nr_genes}{???}

\item{gene_i}{???}

\item{gene_j}{???}
}
\description{
Functional similarity between two genes in annotation database
}
