% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{graph_from_op}
\alias{graph_from_op}
\title{Creation of igraph object from Omnipath dataset}
\usage{
graph_from_op(op_data)
}
\arguments{
\item{op_data}{???}
}
\value{
???
}
\description{
Creation of igraph object from Omnipath dataset
}
