% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{graph_to_adjacency}
\alias{graph_to_adjacency}
\title{Convert network graph into adjacency matrix}
\usage{
graph_to_adjacency(graph_op)
}
\arguments{
\item{graph_op}{???}
}
\value{
???
}
\description{
Convert network graph into adjacency matrix
}
