% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{isgene_omnipath}
\alias{isgene_omnipath}
\title{Check which genes are or not in Omnipath}
\usage{
isgene_omnipath(graph_op, gene_set, exist_bol)
}
\arguments{
\item{graph_op}{???}

\item{gene_set}{???}

\item{exist_bol}{???}
}
\value{
???
}
\description{
Check which genes are or not in Omnipath
}
