% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{prioritization_genes}
\alias{prioritization_genes}
\title{Prioritization of genes based on probabilities}
\usage{
prioritization_genes(
  graph_op,
  prob_matrix,
  genes_interest,
  percentage_genes_ranked
)
}
\arguments{
\item{graph_op}{???}

\item{prob_matrix}{???}

\item{genes_interest}{???}

\item{percentage_genes_ranked}{???}
}
\value{
???
}
\description{
Prioritization of genes based on probabilities
}
