% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{random_walk}
\alias{random_walk}
\title{Random Walk on weighted adjacency matrix}
\usage{
random_walk(weighted_adj_matrix, restart_prob, threshold)
}
\arguments{
\item{weighted_adj_matrix}{???}

\item{restart_prob}{???}

\item{threshold}{???}
}
\value{
???
}
\description{
Random Walk on weighted adjacency matrix
}
