% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_app.R
\name{score_candidate_genes_from_PPI}
\alias{score_candidate_genes_from_PPI}
\title{Weight protein-protein interactions}
\usage{
score_candidate_genes_from_PPI(
  genes_interest = NULL,
  HPO_interest = NULL,
  percentage_output_genes = NULL,
  graph_order = NULL,
  GO_annot = TRUE,
  HPO_annot = TRUE,
  restart_prob_rw = NULL,
  threshold_rw = NULL
)
}
\arguments{
\item{genes_interest}{???}

\item{HPO_interest}{???}

\item{percentage_output_genes}{???}

\item{graph_order}{???}
}
\value{
Data frame with protein-protein interactions of the genes of
     interest scored in the context of the ontology terms of interest.
}
\description{
General function to run all auxiliar functions in WPPI_functions and
Annot_functions
}
