% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{visualize_graph}
\alias{visualize_graph}
\title{Visualize the sub graph}
\usage{
visualize_graph(
  sub_graph,
  genes_interest,
  scores,
  palette = NULL,
  colors = NULL
)
}
\arguments{
\item{sub_graph}{Igraph graph object from \code{\link{subgraph_op}}.}

\item{genes_interest}{Character vector of gene symbols with genes known to
be related to the investigated disease or condition.}

\item{scores}{Data frame created by \code{\link{score_candidate_genes_from_PPI}}
or \code{\link{prioritization_genes}}.}

\item{palette}{An optional character string. Choices include "blue", "green",
"red" or "orange".}

\item{colors}{An optional named list of character vectors, containing colors
for "nodes", "edges", and "genes_interest". Colors for "nodes" and "edges"
are passed on to \code{grDevices::colorRampPalette}, color for "genes_interest"
is passed on to \code{visNetwork::visNetwork}. If both \code{palette} and
\code{colors} are provided, \code{colors} will be prioritized.}
}
\value{
A \code{visNetwork} object visualizing the PPI network of given
genes of interest and their x-order degree neighbors.
}
\description{
Creates a network based on a sub graph, the genes of interest and the
calculated scores. Nodes carry a tooltip listing the gene, its protein code,
and its three neighbors with the highest score.
}
\seealso{
\itemize{
    \item{\code{\link{subgraph_op}}}
    \item{\code{\link{prioritization_genes}}}
    \item{\code{\link{score_candidate_genes_from_PPI}}}
}
}
