% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{weighted_adj}
\alias{weighted_adj}
\title{Convert adjacency to weighted adjacency}
\usage{
weighted_adj(graph_op, neighbors_data, GO_data, HPO_data, nr_GO, nr_HPO)
}
\arguments{
\item{graph_op}{???}

\item{neighbors_data}{???}

\item{GO_data}{???}

\item{HPO_data}{???}

\item{nr_GO}{???}

\item{nr_HPO}{???}
}
\value{
???
}
\description{
Convert adjacency to weighted adjacency
}
