% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_data}
\alias{wppi_data}
\title{Compile data for WPPI}
\usage{
wppi_data()
}
\value{
Returns `NULL`.
}
\description{
This I will largely rewrite, for now it's fine as it is, I just wanted to
have the docstring for roxygen.
}
