% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_go_data}
\alias{wppi_go_data}
\title{Retrieves data from Gene Ontology (HPO)}
\usage{
wppi_go_data()
}
\value{
A data frame (tibble) with GO data.
}
\description{
Gene Ontology (\url{http://geneontology.org/}), GO) annotates genes
by their function, localization and biological processes.
}
\examples{
go <- wppi_go_data()

}
\seealso{
\code{\link{wppi_data}}
}
