% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_hpo_data}
\alias{wppi_hpo_data}
\title{Retrieves data from Human Phenotype Ontology (HPO)}
\usage{
wppi_hpo_data()
}
\value{
A data frame (tibble) with HPO data.
}
\description{
Human Phenotype Ontology (\url{https://hpo.jax.org/app/}), HPO) annotates
proteins with phenotypes and diseases.
}
\examples{
hpo <- wppi_hpo_data()

}
\seealso{
\code{\link{wppi_data}}
}
