% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_omnipath_data}
\alias{wppi_omnipath_data}
\title{Protein-protein interaction data from OmniPath}
\usage{
wppi_omnipath_data(...)
}
\arguments{
\item{...}{Passed to
\code{OmnipathR::import_post_translational_interactions}.}
}
\value{
A data frame (tibble) with protein-protein interaction data from
    OmniPath.
}
\description{
OmniPath (\url{https://omnipathdb.org/}) integrates protein-protein
interactions (PPI) from more than 30 resources. The network created is 
highly customizable by passing parameters to
\code{OmnipathR::import_post_translational_interactions}.
}
\examples{
omnipath <- wppi_omnipath_data()

}
\seealso{
\code{\link{wppi_data}}
}
