% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_uniprot_data}
\alias{wppi_uniprot_data}
\title{Retrieve data from UniProt}
\usage{
wppi_uniprot_data()
}
\value{
A data frame (tibble) with GO data.
}
\description{
UniProt (\url{https://uniprot.org/}) serves as a reference database for
the human proteome and provides the primary identifier for proteins used
in this package.
}
\examples{
uniprot <- wppi_uniprot_data()

}
\seealso{
\code{\link{wppi_data}}
}
