% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_functions.R
\name{functional_annot}
\alias{functional_annot}
\title{Functional similarity between two genes in ontology database (GO or HPO). For
each pair of interacting proteins in the PPI graph network, is quantified the
shared annotations between them using the Fisher's combined probability test
(\url{https://doi.org/10.1007/978-1-4612-4380-9_6}). This is based on the 
number of genes annotated in each shared ontology term and the total amount 
of unique genes available in the ontology database.}
\usage{
functional_annot(data_aggregated, nr_genes, gene_i, gene_j)
}
\arguments{
\item{data_aggregated}{Data frame with gene symbols aggregated by annotation 
for GO/HPO databases.}

\item{nr_genes}{Integer value with number of total unique genes in ontology 
database.}

\item{gene_i}{String with the gene symbol in the row of the adjacency matrix.}

\item{gene_j}{String with the gene symbol in the column of the adjacency 
matrix.}
}
\value{
Integer value with GO/HPO functional similarity between given pair of 
proteins.
}
\description{
Functional similarity between two genes in ontology database (GO or HPO). For
each pair of interacting proteins in the PPI graph network, is quantified the
shared annotations between them using the Fisher's combined probability test
(\url{https://doi.org/10.1007/978-1-4612-4380-9_6}). This is based on the 
number of genes annotated in each shared ontology term and the total amount 
of unique genes available in the ontology database.
}
