% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{graph_to_adjacency}
\alias{graph_to_adjacency}
\title{Convert network graph into adjacency matrix}
\usage{
graph_to_adjacency(graph_op)
}
\arguments{
\item{graph_op}{Igraph graph object.}
}
\value{
Matrix: the adjacency matrix of the graph.
}
\description{
Convert network graph into adjacency matrix
}
\examples{
\donttest{
interactions <- OmnipathR::import_omnipath_interactions()
graph <- OmnipathR::interaction_graph(interactions)
adj_matrix <- graph_to_adjacency(graph)
}

}
