% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_app.R
\name{score_candidate_genes_from_PPI}
\alias{score_candidate_genes_from_PPI}
\title{The full WPPI workflow}
\usage{
score_candidate_genes_from_PPI(
  genes_interest,
  HPO_interest = NULL,
  percentage_output_genes = 100,
  graph_order = 1,
  GO_annot = TRUE,
  HPO_annot = TRUE,
  restart_prob_rw = 0.4,
  threshold_rw = 1e-05,
  databases = NULL
)
}
\arguments{
\item{genes_interest}{Character vector with known-disease specific genes.}

\item{HPO_interest}{Character vector with Human Phenotype Ontology (HPO)
annotations of interest from which to construct the functionality (for
a list of available annotations call \code{\link{wppi_data}}). If not
specified, all the annotations available in the HPO database will be
used.}

\item{percentage_output_genes}{Positive integer (range between 0 and 100)
specifying the percentage (%) of the total candidate genes in the
network returned in the output. If not specified, the score of all the
candidate genes is delivered.}

\item{graph_order}{Positive integer bigger than 0 which defines the x-order
neighbors of the given genes of interest. These new genes, also called
candidate genes, together with the given genes of interest define the
Protein-Protein Interaction (PPI) network used in the analysis. If not
specified, is used the first-order neighbors.}

\item{GO_annot}{Boolean parameter declaring to use or not the Gene Ontology
(GO) annotation database to weight the PPI network. The default setting
is to use it (GO_annot = TRUE).}

\item{HPO_annot}{Boolean parameter declaring to use or not the Human
Phenotype Ontology (HPO) annotation database to weight the PPI network.
The default setting is to use it (HPO_annot = TRUE).}

\item{restart_prob_rw}{Positive value between 0 and 1 defining the restart
probability parameter used in the Random Walk with Restart algorithm.
If not specified, 0.4 is the default value.}

\item{threshold_rw}{Positive value depicting the threshold parameter in the
Random Walk with Restart algorithm. When the error between
probabilities is smaller than the threshold defined, the algorithm
stops. If not specified, 10^(-6) is the default value.}

\item{databases}{Database knowledge as produced by \code{\link{wppi_data}}.}
}
\value{
Data frame with the ranked candidate genes based on the functional
    score inferred from given ontology terms, PPI and Random Walk with
   Restart parameters.
}
\description{
The wppi package is a functional prioritization of new disease specific
genes based on a given set of known disease-related genes and annotation
Weighted Protein-Protein Interactions (WPPI). The PPI networks are obtained
using the Omnipath (\url{https://omnipathdb.org/}) resource and
functionality is deduced using the Gene Ontology (GO,
\url{http://geneontology.org/}) and Human Phenotype Ontology (HPO,
\url{https://hpo.jax.org/app/}) ontology databases. To score the candidate
genes, a Random Walk with Restart algorithm is applied on the weighted
network.
}
\examples{
# example gene set
genes_interest <-
    c("ERCC8", "AKT3", "NOL3", "GFI1B", "CDC25A", "TPX2", "SHE")
# example HPO annotations set
hpo <- wppi_hpo_data()
HPO_interest <- unique(
    dplyr::filter(hpo, grepl("Diabetes", .data$Name))$Name
)
# Score 1st-order candidate genes
new_genes_diabetes <-
    score_candidate_genes_from_PPI(
        genes_interest = genes_interest,
        HPO_interest = HPO_interest,
        percentage_output_genes = 10,
        graph_order = 1)

}
\seealso{
\itemize{
    \item{\code{\link{wppi_data}}}
    \item{\code{\link{weighted_adj}}}
    \item{\code{\link{random_walk}}}
    \item{\code{\link{prioritization_genes}}}
}
}
