% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{subgraph_op}
\alias{subgraph_op}
\title{Subgraph from Omnipath graph object and genes of interest}
\usage{
subgraph_op(graph_op, gene_set, sub_level)
}
\arguments{
\item{graph_op}{???}

\item{gene_set}{???}

\item{sub_level}{???}
}
\value{
???
}
\description{
Subgraph from Omnipath graph object and genes of interest
}
