% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_functions.R
\name{weighted_adj}
\alias{weighted_adj}
\title{Weighted adjacency matrix}
\usage{
weighted_adj(graph_op, neighbors_data, GO_data, HPO_data)
}
\arguments{
\item{graph_op}{Igraph object based on Omnipath PPI interactions from
\code{\link{graph_from_op}}.}

\item{neighbors_data}{Data table output from
\code{\link{functional_annot}}.}

\item{GO_data}{Data frame with GO annotations filtered and aggregated for
the proteins/genes available in the graph object.}

\item{HPO_data}{Data frame with HPO annotations filtered and aggregated
for the proteins/genes available in the graph object.}
}
\value{
Weighted adjacency matrix based on network topology and functional
    similarity between interacting proteins/genes based on ontology
    databases.
}
\description{
Converts adjacency to weighted adjacency using network topology
information (shared neighbors between connected nodes via
\code{\link{common_neighbors}}) integrated with genome and phenotype
factors from GO and HPO annotation terms (functionality computed by
\code{\link{functional_annot}}). At the end, the weighted adjacency
matrix is normalized by column.
}
\examples{
db <- wppi_data()
genes_interest <-
    c("ERCC8", "AKT3", "NOL3", "GFI1B", "CDC25A", "TPX2", "SHE")
graph_op <- graph_from_op(db$omnipath)
graph_op_1 <- subgraph_op(graph_op, genes_interest, 1)
neighbors_data <- common_neighbors(graph_op_1)
w_adj <- weighted_adj(graph_op_1, neighbors_data, db$go, db$hpo)

}
