% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_data}
\alias{wppi_data}
\title{Database knowledge for wppi}
\usage{
wppi_data(...)
}
\arguments{
\item{...}{Passed to
\code{OmnipathR::import_post_translational_interactions}.}
}
\value{
A list of data frames (tibbles) with database knowledge from HPO,
    GO, OmniPath and UniProt.
}
\description{
Retrieves the database knowledge necessary for WPPI directly from the
databases. The databases used here are the Human Phenotype Ontology (HPO,
\url{https://hpo.jax.org/app/}), Gene Ontology (GO,
\url{http://geneontology.org/}), OmniPath (\url{https://omnipathdb.org/})
and UniProt (\url{https://uniprot.org/}). The downloads carried out by
the OmnipathR package and data required by wppi are extracted from each
table.
}
\examples{
# Download all data
data_wppi <- wppi_data()
# Omnipath
omnipath_data <- data_wppi$omnipath
# HPO
HPO_data <- data_wppi$hpo
# GO
GO_data <- data_wppi$go

}
\seealso{
\itemize{
    \item{\code{\link{wppi_go_data}}}
    \item{\code{\link{wppi_hpo_data}}}
    \item{\code{\link{wppi_omnipath_data}}}
    \item{\code{\link{wppi_uniprot_data}}}
}
}
