% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPPI_data.R
\name{wppi_data}
\alias{wppi_data}
\title{Compile database knowledge for WPPI}
\usage{
wppi_data(...)
}
\arguments{
\item{...}{Passed to
\code{OmnipathR::import_post_translational_interactions}.}
}
\value{
A list of data frames (tibbles) with database knowledge from HPO,
    GO, OmniPath and UniProt.
}
\description{
Retrieves the database knowledge necessary for WPPI directly from the
databases. The databases used here are the Human Phenotype Ontology (HPO,
\url{https://hpo.jax.org/app/}), Gene Ontology (GO,
\url{http://geneontology.org/}), OmniPath (\url{https://omnipathdb.org/})
and UniProt (\url{https://uniprot.org/}). The downloads carried out by
the OmnipathR package and data required by WPPI is extracted from each
table.
}
